worked on access & refress token

+++ frontend/src/app/api/tasks/[id]/route.ts
@@ -7,6 +7,14 @@ import { verifyJwtToken } from '../../middleware/checkAuth';
 
 export async function GET(request: NextRequest, { params }) {
     try {
+
+        const authorization = request.headers.get('Authorization');
+        if (!authorization || !authorization.startsWith('Bearer ')) {
+            return new NextResponse('Unauthorized: Bearer token missing or invalid', { status: 401 });
+        }
+
+        const accressToken = authorization.split(' ')[1];
+        await verifyJwtToken(accressToken);
         const { id } = await params;
         const Task = await getTasks(id);
         return NextResponse.json(Task, { status: 200 });
@@ -41,7 +49,13 @@ export async function DELETE(request: NextRequest, { params }) {
             return NextResponse.json({ msg: "Invalid ID format" }, { status: 400 });
         }
 
-        await verifyJwtToken();
+         const authorization = request.headers.get('Authorization');
+        if (!authorization || !authorization.startsWith('Bearer ')) {
+            return new NextResponse('Unauthorized: Bearer token missing or invalid', { status: 401 });
+        }
+
+        const accressToken = authorization.split(' ')[1];
+        await verifyJwtToken(accressToken);
 
         await deleteTasks(id);
         return NextResponse.json({ msg: 'Task Deleted' }, { status: 200 })
